spline_evaluator_batched -> spline_evaluator
