+++ include/ddc/kernels/splines/spline_evaluator.hpp
@@ -1,16 +1,22 @@
 #pragma once
 
 #include <array>
-#include <cmath>
 
 #include <ddc/ddc.hpp>
 
+#include "Kokkos_Macros.hpp"
 #include "spline_boundary_value.hpp"
 #include "view.hpp"
 
 namespace ddc {
-template <class ExecSpace, class MemorySpace, class BSplinesType, class interpolation_mesh_type>
-class SplineEvaluator
+
+template <
+        class ExecSpace,
+        class MemorySpace,
+        class BSplinesType,
+        class InterpolationMesh,
+        class... IDimX>
+class SplineEvaluatorBatched
 {
 private:
     // Tags to determine what to evaluate
@@ -22,6 +28,8 @@ private:
     {
     };
 
+    using tag_type = typename BSplinesType::tag_type;
+
 public:
     using exec_space = ExecSpace;
 
@@ -29,180 +37,211 @@ public:
 
     using bsplines_type = BSplinesType;
 
-    using tag_type = typename BSplinesType::tag_type;
+    using interpolation_mesh_type = InterpolationMesh;
 
-    using mesh_type = interpolation_mesh_type;
+    using interpolation_domain_type = ddc::DiscreteDomain<interpolation_mesh_type>;
 
-    using interpolation_domain_type = ddc::DiscreteDomain<interpolation_mesh_type>; // Use ?
+    using vals_domain_type = ddc::DiscreteDomain<IDimX...>;
+
+    using bsplines_domain_type = ddc::DiscreteDomain<bsplines_type>;
+
+    using batch_domain_type =
+            typename ddc::detail::convert_type_seq_to_discrete_domain<ddc::type_seq_remove_t<
+                    ddc::detail::TypeSeq<IDimX...>,
+                    ddc::detail::TypeSeq<interpolation_mesh_type>>>;
+
+    template <typename Tag>
+    using spline_dim_type
+            = std::conditional_t<std::is_same_v<Tag, interpolation_mesh_type>, bsplines_type, Tag>;
+
+    using spline_domain_type =
+            typename ddc::detail::convert_type_seq_to_discrete_domain<ddc::type_seq_replace_t<
+                    ddc::detail::TypeSeq<IDimX...>,
+                    ddc::detail::TypeSeq<interpolation_mesh_type>,
+                    ddc::detail::TypeSeq<bsplines_type>>>;
+
 
 private:
-    SplineBoundaryValue<BSplinesType> const& m_left_bc;
+    const spline_domain_type m_spline_domain;
 
-    SplineBoundaryValue<BSplinesType> const& m_right_bc;
 
 public:
-    SplineEvaluator() = delete;
-
-    explicit SplineEvaluator(
-            SplineBoundaryValue<BSplinesType> const& left_bc,
-            SplineBoundaryValue<BSplinesType> const& right_bc)
-        : m_left_bc(left_bc)
-        , m_right_bc(right_bc)
+    explicit SplineEvaluatorBatched(
+            spline_domain_type const& spline_domain,
+            SplineBoundaryValue<bsplines_type> const& left_bc, // Unused, to be restored in next MR
+            SplineBoundaryValue<bsplines_type> const& right_bc)
+        : m_spline_domain(spline_domain)
     {
     }
 
-    SplineEvaluator(SplineEvaluator const& x) = default;
+    SplineEvaluatorBatched(SplineEvaluatorBatched const& x) = default;
 
-    SplineEvaluator(SplineEvaluator&& x) = default;
+    SplineEvaluatorBatched(SplineEvaluatorBatched&& x) = default;
 
-    ~SplineEvaluator() = default;
+    ~SplineEvaluatorBatched() = default;
 
-    SplineEvaluator& operator=(SplineEvaluator const& x) = default;
+    SplineEvaluatorBatched& operator=(SplineEvaluatorBatched const& x) = default;
 
-    SplineEvaluator& operator=(SplineEvaluator&& x) = default;
+    SplineEvaluatorBatched& operator=(SplineEvaluatorBatched&& x) = default;
 
-    /*
-	SplineBoundaryValue<bsplines_type> left_bc() const noexcept
-        {
-            return m_left_bc;
-        }
 
-        SplineBoundaryValue<bsplines_type> right_bc() const noexcept
-        {
-          return m_right_bc;
-        }
-	*/
 
-    double operator()(
-            ddc::Coordinate<tag_type> const& coord_eval,
-            ddc::ChunkSpan<double const, ddc::DiscreteDomain<BSplinesType>> const spline_coef) const
+    KOKKOS_FUNCTION spline_domain_type spline_domain() const noexcept
     {
-        std::array<double, bsplines_type::degree() + 1> values;
-        ddc::DSpan1D const vals = as_span(values);
-
-        return eval(coord_eval, spline_coef, vals);
+        return m_spline_domain;
     }
 
-    template <class Domain, class Layout1, class Layout2, class Layout3>
+    KOKKOS_FUNCTION bsplines_domain_type bsplines_domain() const noexcept // TODO : clarify name
+    {
+        return ddc::discrete_space<bsplines_type>().full_domain();
+    }
+
+    KOKKOS_FUNCTION batch_domain_type batch_domain() const noexcept
+    {
+        return ddc::remove_dims_of(spline_domain(), bsplines_domain());
+    }
+
+    template <class Layout, class... CoordsDims>
+    KOKKOS_FUNCTION double operator()(
+            ddc::Coordinate<CoordsDims...> const& coord_eval,
+            ddc::ChunkSpan<double const, spline_domain_type, Layout, memory_space> const
+                    spline_coef) const
+    {
+        return eval(coord_eval, spline_coef);
+    }
+
+    template <class Layout1, class Layout2, class Layout3, class... CoordsDims>
     void operator()(
-            ddc::ChunkSpan<double, Domain, Layout1, MemorySpace> const spline_eval,
-            ddc::ChunkSpan<const ddc::Coordinate<tag_type>, Domain, Layout2, MemorySpace> const
-                    coords_eval,
+            ddc::ChunkSpan<double, vals_domain_type, Layout1, memory_space> const spline_eval,
             ddc::ChunkSpan<
-                    double const,
-                    ddc::DiscreteDomain<BSplinesType>,
-                    Layout3,
-                    MemorySpace> const spline_coef) const
+                    ddc::Coordinate<CoordsDims...> const,
+                    vals_domain_type,
+                    Layout2,
+                    memory_space> const coords_eval,
+            ddc::ChunkSpan<double const, spline_domain_type, Layout3, memory_space> const
+                    spline_coef) const
     {
-        for (auto i : coords_eval.domain()) {
-            spline_eval(i) = eval(coords_eval(i), spline_coef);
-        }
+        interpolation_domain_type const interpolation_domain(spline_eval.domain());
+        ddc::for_each(
+                ddc::policies::policy(exec_space()),
+                batch_domain(),
+                KOKKOS_CLASS_LAMBDA(typename batch_domain_type::discrete_element_type const j) {
+                    const auto spline_eval_1D = spline_eval[j];
+                    const auto coords_eval_1D = coords_eval[j];
+                    const auto spline_coef_1D = spline_coef[j];
+                    for (auto const i : interpolation_domain) {
+                        spline_eval_1D(i) = eval(coords_eval_1D(i), spline_coef_1D);
+                    }
+                });
     }
 
-    template <class Layout>
-    double deriv(
-            ddc::Coordinate<tag_type> const& coord_eval,
-            ddc::ChunkSpan<
-                    double const,
-                    ddc::DiscreteDomain<BSplinesType>,
-                    Layout,
-                    MemorySpace> const spline_coef) const
+    template <class Layout, class... CoordsDims>
+    KOKKOS_FUNCTION double deriv(
+            ddc::Coordinate<CoordsDims...> const& coord_eval,
+            ddc::ChunkSpan<double const, bsplines_domain_type, Layout, memory_space> const
+                    spline_coef) const
     {
-        std::array<double, bsplines_type::degree() + 1> values;
-        ddc::DSpan1D const vals = as_span(values);
-
-        return eval_no_bc(coord_eval, spline_coef, vals, eval_deriv_type());
+        return eval_no_bc<eval_deriv_type>(coord_eval, spline_coef);
     }
 
-    template <class Domain, class Layout1, class Layout2, class Layout3>
+    template <class Layout1, class Layout2, class Layout3, class... CoordsDims>
     void deriv(
-            ddc::ChunkSpan<double, Domain, Layout1, MemorySpace> const spline_eval,
-            ddc::ChunkSpan<const ddc::Coordinate<tag_type>, Domain, Layout2, MemorySpace> const
-                    coords_eval,
+            ddc::ChunkSpan<double, vals_domain_type, Layout1, memory_space> const spline_eval,
             ddc::ChunkSpan<
-                    double const,
-                    ddc::DiscreteDomain<BSplinesType>,
-                    Layout3,
-                    MemorySpace> const spline_coef) const
+                    ddc::Coordinate<CoordsDims...> const,
+                    vals_domain_type,
+                    Layout2,
+                    memory_space> const coords_eval,
+            ddc::ChunkSpan<double const, spline_domain_type, Layout3, memory_space> const
+                    spline_coef) const
     {
-        for (auto i : coords_eval.domain()) {
-            spline_eval(i) = eval_no_bc(coords_eval(i), spline_coef, eval_deriv_type());
-        }
+        interpolation_domain_type const interpolation_domain(spline_eval.domain());
+        ddc::for_each(
+                ddc::policies::policy(exec_space()),
+                batch_domain(),
+                KOKKOS_CLASS_LAMBDA(typename batch_domain_type::discrete_element_type const j) {
+                    const auto spline_eval_1D = spline_eval[j];
+                    const auto coords_eval_1D = coords_eval[j];
+                    const auto spline_coef_1D = spline_coef[j];
+                    for (auto const i : interpolation_domain) {
+                        spline_eval_1D(i)
+                                = eval_no_bc<eval_deriv_type>(coords_eval_1D(i), spline_coef_1D);
+                    }
+                });
     }
 
-    template <class Layout>
-    double integrate(ddc::ChunkSpan<
-                     double const,
-                     ddc::DiscreteDomain<BSplinesType>,
-                     Layout,
-                     MemorySpace> const spline_coef) const
+    template <class Layout1, class Layout2>
+    void integrate(
+            ddc::ChunkSpan<double, batch_domain_type, Layout1, memory_space> const integrals,
+            ddc::ChunkSpan<double const, spline_domain_type, Layout2, memory_space> const
+                    spline_coef) const
     {
-        ddc::Chunk values(spline_coef.domain(), ddc::KokkosAllocator<double, MemorySpace>());
+        ddc::Chunk values_alloc(
+                ddc::DiscreteDomain<bsplines_type>(spline_coef.domain()),
+                ddc::KokkosAllocator<double, memory_space>());
+        ddc::ChunkSpan values = values_alloc.span_view();
+        Kokkos::parallel_for(
+                Kokkos::RangePolicy<exec_space>(0, 1),
+                KOKKOS_LAMBDA(int) { ddc::discrete_space<bsplines_type>().integrals(values); });
 
-        ddc::discrete_space<bsplines_type>().integrals(values.span_view());
-
-        return ddc::transform_reduce(
-                spline_coef.domain(),
-                0.0,
-                ddc::reducer::sum<double>(),
-                [&](ddc::DiscreteElement<BSplinesType> const ibspl) {
-                    return spline_coef(ibspl) * values(ibspl);
+        ddc::for_each(
+                ddc::policies::policy(exec_space()),
+                batch_domain(),
+                KOKKOS_LAMBDA(typename batch_domain_type::discrete_element_type const j) {
+                    integrals(j) = 0;
+                    for (typename bsplines_domain_type::discrete_element_type const i :
+                         values.domain()) {
+                        integrals(j) += spline_coef(i, j) * values(i);
+                    }
                 });
     }
 
 private:
-    template <class Layout>
-    double eval(
-            ddc::Coordinate<tag_type> coord_eval,
-            ddc::ChunkSpan<
-                    double const,
-                    ddc::DiscreteDomain<BSplinesType>,
-                    Layout,
-                    MemorySpace> const spline_coef) const
+    template <class Layout, class... CoordsDims>
+    KOKKOS_INLINE_FUNCTION double eval(
+            ddc::Coordinate<CoordsDims...> const& coord_eval,
+            ddc::ChunkSpan<double const, bsplines_domain_type, Layout, memory_space> const
+                    spline_coef) const
     {
+        ddc::Coordinate<typename interpolation_mesh_type::continuous_dimension_type>
+                coord_eval_interpolation
+                = ddc::select<typename interpolation_mesh_type::continuous_dimension_type>(
+                        coord_eval);
         if constexpr (bsplines_type::is_periodic()) {
-            if (coord_eval < ddc::discrete_space<bsplines_type>().rmin()
-                || coord_eval > ddc::discrete_space<bsplines_type>().rmax()) {
-                coord_eval -= std::floor(
-                                      (coord_eval - ddc::discrete_space<bsplines_type>().rmin())
-                                      / ddc::discrete_space<bsplines_type>().length())
-                              * ddc::discrete_space<bsplines_type>().length();
-            }
-        } else {
-            if (coord_eval < ddc::discrete_space<bsplines_type>().rmin()) {
-                return m_left_bc(coord_eval, spline_coef);
-            }
-            if (coord_eval > ddc::discrete_space<bsplines_type>().rmax()) {
-                return m_right_bc(coord_eval, spline_coef);
+            if (coord_eval_interpolation < ddc::discrete_space<bsplines_type>().rmin()
+                || coord_eval_interpolation > ddc::discrete_space<bsplines_type>().rmax()) {
+                coord_eval_interpolation -= Kokkos::floor(
+                                                    (coord_eval_interpolation
+                                                     - ddc::discrete_space<bsplines_type>().rmin())
+                                                    / ddc::discrete_space<bsplines_type>().length())
+                                            * ddc::discrete_space<bsplines_type>().length();
             }
         }
-        return eval_no_bc(coord_eval, spline_coef, eval_type());
+        return eval_no_bc<eval_type>(coord_eval_interpolation, spline_coef);
     }
 
-    template <class EvalType, class Layout>
-    double eval_no_bc(
-            ddc::Coordinate<tag_type> const& coord_eval,
-            ddc::ChunkSpan<
-                    double const,
-                    ddc::DiscreteDomain<BSplinesType>,
-                    Layout,
-                    MemorySpace> const spline_coef,
-            EvalType const) const
+    template <class EvalType, class Layout, class... CoordsDims>
+    KOKKOS_INLINE_FUNCTION double eval_no_bc(
+            ddc::Coordinate<CoordsDims...> const& coord_eval,
+            ddc::ChunkSpan<double const, bsplines_domain_type, Layout, memory_space> const
+                    spline_coef) const
     {
         static_assert(
                 std::is_same_v<EvalType, eval_type> || std::is_same_v<EvalType, eval_deriv_type>);
-        ddc::DiscreteElement<BSplinesType> jmin;
-
+        ddc::DiscreteElement<bsplines_type> jmin;
         std::array<double, bsplines_type::degree() + 1> vals;
+        ddc::Coordinate<typename interpolation_mesh_type::continuous_dimension_type>
+                coord_eval_interpolation
+                = ddc::select<typename interpolation_mesh_type::continuous_dimension_type>(
+                        coord_eval);
         if constexpr (std::is_same_v<EvalType, eval_type>) {
-            jmin = ddc::discrete_space<bsplines_type>().eval_basis(vals, coord_eval);
+            jmin = ddc::discrete_space<bsplines_type>().eval_basis(vals, coord_eval_interpolation);
         } else if constexpr (std::is_same_v<EvalType, eval_deriv_type>) {
-            jmin = ddc::discrete_space<bsplines_type>().eval_deriv(vals, coord_eval);
+            jmin = ddc::discrete_space<bsplines_type>().eval_deriv(vals, coord_eval_interpolation);
         }
-
         double y = 0.0;
         for (std::size_t i = 0; i < bsplines_type::degree() + 1; ++i) {
-            y += spline_coef(jmin + i) * vals[i];
+            y += spline_coef(ddc::DiscreteElement<bsplines_type>(jmin + i)) * vals[i];
         }
         return y;
     }
